docs: document parser registry
