+++ .changeset/refactor-parser-registry.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': patch
+---
+
+refactor parser into language-specific modules and add registry-based dispatch
