+++ src/core/linter.ts
@@ -1,15 +1,21 @@
 import pLimit from 'p-limit';
 import os from 'node:os';
-import type { LintResult, DesignTokens } from './types.js';
-import { normalizeTokens } from './token-loader.js';
+import type {
+  LintResult,
+  DesignTokens,
+  LintMessage,
+  RuleContext,
+} from './types.js';
+import { normalizeTokens, mergeTokens } from './token-loader.js';
 import { extractVarName } from '../utils/token-match.js';
 export { defaultIgnore } from './ignore.js';
 import type { Cache } from './cache.js';
 import { RuleRegistry } from './rule-registry.js';
-import { ParserService } from './parser-service.js';
 import { TokenTracker } from './token-tracker.js';
 import { FileService } from './file-service.js';
 import { CacheService } from './cache-service.js';
+import { parserRegistry } from './parser-registry.js';
+import path from 'node:path';
 
 export interface Config {
   tokens?: DesignTokens | Record<string, DesignTokens>;
@@ -33,7 +39,6 @@ export class Linter {
   private config: ResolvedConfig;
   private tokensByTheme: Record<string, DesignTokens> = {};
   private ruleRegistry: RuleRegistry;
-  private parser: ParserService;
   private tokenTracker: TokenTracker;
 
   constructor(config: Config) {
@@ -45,7 +50,6 @@ export class Linter {
     this.config = { ...config, tokens: normalized.merged };
     this.ruleRegistry = new RuleRegistry(this.config);
     this.tokenTracker = new TokenTracker(this.config.tokens);
-    this.parser = new ParserService(this.tokensByTheme);
   }
 
   async lintFile(
@@ -144,8 +148,67 @@ export class Linter {
     if (this.tokenTracker.hasUnusedTokenRules()) {
       this.tokenTracker.trackUsage(text);
     }
-    return this.parser.lintText(text, filePath, enabled);
+    const messages: LintMessage[] = [];
+    const ruleDescriptions: Record<string, string> = {};
+    const disabledLines = getDisabledLines(text);
+    const listeners = enabled.map(({ rule, options, severity }) => {
+      ruleDescriptions[rule.name] = rule.meta.description;
+      const themes =
+        isRecord(options) && isStringArray(options.themes)
+          ? options.themes
+          : undefined;
+      const tokens = mergeTokens(this.tokensByTheme, themes);
+      const ctx: RuleContext = {
+        filePath,
+        tokens,
+        options,
+        report: (m) => messages.push({ ...m, severity, ruleId: rule.name }),
+      };
+      return rule.create(ctx);
+    });
+    const ext = path.extname(filePath).toLowerCase();
+    const parser = parserRegistry[ext];
+    if (parser) {
+      await parser(text, filePath, listeners, messages);
+    }
+    const filtered = messages.filter((m) => !disabledLines.has(m.line));
+    return { filePath, messages: filtered, ruleDescriptions };
   }
 }
 
 export { applyFixes } from './cache-manager.js';
+
+function getDisabledLines(text: string): Set<number> {
+  const disabled = new Set<number>();
+  const lines = text.split(/\r?\n/);
+  let block = false;
+  for (let i = 0; i < lines.length; i++) {
+    const line = lines[i];
+    if (/\/\*\s*design-lint-disable\s*\*\//.test(line)) {
+      block = true;
+      continue;
+    }
+    if (/\/\*\s*design-lint-enable\s*\*\//.test(line)) {
+      block = false;
+      continue;
+    }
+    if (/(?:\/\/|\/\*)\s*design-lint-disable-next-line/.test(line)) {
+      disabled.add(i + 2);
+      continue;
+    }
+    if (line.includes('design-lint-disable-line')) {
+      disabled.add(i + 1);
+      continue;
+    }
+    if (block) disabled.add(i + 1);
+  }
+  return disabled;
+}
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === 'object' && value !== null;
+}
+
+function isStringArray(value: unknown): value is string[] {
+  return Array.isArray(value) && value.every((v) => typeof v === 'string');
+}
